% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.lg.R
\name{fst.lg}
\alias{fst.lg}
\title{Calculate Fst per linkage group}
\usage{
fst.lg(lg)
}
\arguments{
\item{lg}{Vector of linkage groups}
}
\value{
Value Fst values
}
\description{
This function calculates Hudson's Fst
}
\examples{
fst.stats.lgs<-lapply(lgs, fst.lg)

}
