% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.regions.R
\name{get.regions}
\alias{get.regions}
\title{function to get size of regions}
\usage{
get.regions(lb = NA, ub = NA, loci.ord = NA, loci.pos = NA, lgordscalen = NA)
}
\arguments{
\item{lb}{input}

\item{loci.ord}{input}

\item{loci.pos}{input}

\item{lgordscalen}{input}
}
\value{

}
\description{
function to get size of regions
}
\examples{

}
