% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.normc.R
\name{Mstep.normc}
\alias{Mstep.normc}
\title{Modified version of Mstep.norm for fitting HMM
#'}
\usage{
Mstep.normc(x, cond, pm, pn)
}
\arguments{
\item{x}{input}

\item{cond}{input}

\item{pm}{input}

\item{pn}{input}
}
\value{

}
\description{
Modified version of Mstep.norm for fitting HMM
#'
}
\examples{

}
